+++ simulator/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to the contest simulator will be documented in this file.
 Dates are based on Eastern Time Zone.
 
+## 2024-11-23
+
+### Changed
+
+- Updated first-turn rates of season 13
+
 ## 2024-11-21
 
 ### Added